+++ packages/marketing-components/src/trustelements/SecurityIllustration/SecurityIllustration.less
