+++ packages/marketing-components/src/trustelements/SecurityIllustration/index.js
